CI part 1 (#4)
* Minor cleanup (!!!)

* Starting work on utf8lex continuous integration.

* Fixed .github-ci indent

* My GitHub Actions fu is rusty

* Actually if the missing colon is the problem then my YAML fu is even rustier

* Both of the above?

* More YAML sigh

* No debian runners, only ubuntu (https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories)

* Make sure dependencies are installed before compiling.

* Add sudo to dependency installs

* CI already found a bug.  Woot!

* And another bug...

* Use ref name for artifact upload, instead of just ref (which is a path) nb ref-name is also a path on pull request merge...

* Remove overwrite: true (apparently not [yet] supported)

* Add sloccount (https://dwheeler.com/sloccount/) to build container for counting lines and estimating costs (https://packages.debian.org/bookworm/sloccount)

* Continuous integration experiments work in progress from 2024-04-03

* Let's try switching all build and test steps to inside the container image, see how this goes...  Apart from syntax, might fail on multi-arch builds.

* Fix ${{github.ref.name}} -> ${{github.ref_name}}, docker buildx build (path)

* Figure out where the container image went

* Maybe ./build does not exist

* debian container image doesn't exist for darwin platform

* Ugh had scc and sloccount experiments added to utf8lex container file.  Bleah.  Removed, upgraded debian from 12.2 to 12.6.

* Build container, then build shared libraries.  take 1

* Fix ${{BUILD_DIR}} -> ${{env.BUILD_DIR}}

* Have to pull the container image tar artifact before running it

* GitHub Actions are painful

* One last try to not rely on arcane abstruse GitHub majicks to do standard caching

* No slashes in container names; let's see if the cache works...?

* Try to fix docker caching (AGAIN); figure out why make doesn't work

* These research chemicals are making me feel super weird, man, super weird.  /bin/sh no such file

* GitHub actions/cache/restore fails every time.

* Keep trying to fix stupid GitHub caching jeepers

* Keep trying GitHub Actions is awful

* Well at least we have a cache hit now.  Docker buildx build seems to ignore the cache.  Sigh.

* Missed an if statement

* Let's see if we can use the cache since the artifact seems to have disappeared.

* Use cache for the builder container image, instance of caching docker cache (which docker buildx build refuses to use, for some reason, anyway).

* Why is it not working this time

* GitHub's cache seems to be entirely useless

* Maybe the tar problem was using --output=type=tar instead of type=docker or type=oci

* Missing "id:"

* Fix upload artifact steps referring to relative paths.

* Facepalm

* Fix paths

* Fix paths

* Fix if for gzip step

* Remove extra ${{env.BUILD_DIR}} path from downloaded container image.

* Get rid of the "if" in upload artifact - it needs to run every time, we can only access the artifact from the previous job in the current run.  Also try using path = directory (/tmp/builder/) instead to get artifact upload/download working properly.

* There we go.  Need to mkdir -p ./build.

* Switch from --output=type=oci to --output=type=docker (for now)

* Hmmmm Maybe docker import is the wrong command.  Switch to docker image load.

* debugging outputs

* Docker does not support loading multi-platform images from files (https://stackoverflow.com/questions/72945407/how-do-i-import-and-run-a-multi-platform-oci-image-in-docker-for-macos/73058421#73058421).  So we have to use the matrix to build multiple platforms.

* matrix.platform is linux/amd64 etc; needs to be a directory, not part of the filename.

* Need to mkdir -p the matrix.platform directory.

* matrix/platform typo

* mkdir -p /tmp/builder/ needed .../${{matrix.platform}}; clean up shared object building a bit

* Can't use slashes in an artifact name.  GitHub Actions is so awful.

* Forgot to set the MATRIX_PLATFORM environment variable in the first place.

* Use docker image save, since docker buildx build --output seems to always output something that docker image load cannot handle.

* docker buildx build with docker-container needs --load, otherwise it doesn't output the image anywhere at all (and complains about it, too).

* Add --platform to docker run; more debugging messages.

* Have to make ./build directory and sub-dirs writable for the container user to write out files.

* Also need the parent directory of ./build to be writable, otherwise mkdir -p doesn't work (even though the directory already exists).

* What now still can't mkdir -p ../build

* Add debugging to Makefile to see why permissions problems on build dir

* Whoops this time it was tests/build directory -- because we were doing make all at the toplevel, instead of just make build.  We'll do make unit_tests and make integration_tests as separate jobs.

* During shared library build, do not run unit_tests; add separate job for unit_tests.

* Why cannot stat .so file

* Missing "env." from ${{env.UNIT_TESTS_BUILD_DIR}}

* Makefile variables should be overridable.

* The default environment had CC, CFLAGS, LD, LDFLAGS, and so on set, overriding our use of gcc.

* I don't know how, but even after we unset the environment variables, Make is still getting CC=cc and LD=ld

* Makefiles use ${env-var} not $(make-var).  Need to pass as a make-var with "make make-var=...".

* Make sure unit tests use the tests/build directory that has been chmod'ed; don't make build every time we make unit_tests; get rid of superfluous ${{matrix.platform}} directories.

* During tests, we have to ln -s the libutf8.so and libutf8.so.1 to the version being tested e.g. libutf8.so.1.2.3

* Sigh.  GitHub Actions environment variables apparently cannot refer to each other.

* Typo missed a `

* No such file ld trying to find libutf8lex.so probably bad path to unit_tests need extra ..

* Ugh also the Makefile var is TEST_BUILD_DIR not TESTS_BUILD_DIR

* add debugging to figure out why make in unit_tests directory can't find utf8lex.so

* Don't symlink paths.

* Woohoo!  arm64 and s390x passed unit tests!  But mips64le failed for some reason.  Output slightly better diff from failed test; add integration tests to the workflow.

* Add debugging to unit test to figure out why mips64le and ppc64le fail

* more debugging info

* Interesting.  We were not null-terminating printable_str in the case UTF8LEX_MORE was required.  On some platforms, it just happened to be 0 anyway.  But not on all platforms!

* Remove debugging information from utf8lex_printable_str().

* Remove DEBCONF_FRONTEND (variable does nothing); move build.Dockerfile to builder.Dockerfile and update references to the file.

+++ src/utf8lex_definition_regex.c
@@ -354,7 +354,7 @@ static utf8lex_error_t utf8lex_lex_regex(
   {
     fprintf(stderr,
             "*** utf8lex: pcre2 and utf8proc disagree: pcre2 match_length_bytes = %d vs utf8proc = %d\n",
-            match_length_bytes,
+            (int) match_length_bytes,
             token_loc[UTF8LEX_UNIT_BYTE].length);
   }
